feat: add switch program role functionality and update localization
- Implemented the ability to switch program roles in the Profile component.
- Added a new hook `useAccountSwitch` to handle account switching logic.
- Updated the Login page to utilize the new account switching functionality.
- Enhanced the SwitchAccountDialog to preselect the current tenant and role after login.
- Added localization entries for "SWITCH_PROGRAM_ROLE" in multiple languages.
