+++ apps/admin-app-repo/public/locales/gu/common.json
@@ -94,6 +94,7 @@
     "NO_DISTRICTS": "કોઈ જિલ્લો નથી",
     "ALL_STATES": "તમામ રાજ્યો",
     "LOGOUT": "લૉગ આઉટ કરો",
+    "SWITCH_PROGRAM_ROLE": "એક્સેસ બદલો",
     "HI": "નમસ્તે, {{name}}",
     "MOBILE": "મોબાઈલ",
     "ALL_BLOCKS": "બધા બ્લોક્સ",
