+++ apps/admin-app-repo/public/locales/hi/common.json
@@ -94,6 +94,7 @@
     "NO_DISTRICTS": "कोणतेही जिल्हे नाहीत",
     "ALL_STATES": "सर्व राज्ये",
     "LOGOUT": "लॉग आउट",
+    "SWITCH_PROGRAM_ROLE": "एक्सेस बदलें",
     "HI": "नमस्कार, {{name}}",
     "MOBILE": "मोबाइल",
     "ALL_BLOCKS": "सर्व ब्लॉक्स",
