+++ apps/admin-app-repo/public/locales/kan/common.json
@@ -94,6 +94,7 @@
     "NO_DISTRICTS": "ಜಿಲ್ಲೆಗಳಿಲ್ಲ",
     "ALL_STATES": "ಎಲ್ಲಾ ರಾಜ್ಯಗಳು",
     "LOGOUT": "ಲಾಗ್ ಔಟ್ ಮಾಡಿ",
+    "SWITCH_PROGRAM_ROLE": "ಪ್ರವೇಶ ಬದಲಾಯಿಸಿ",
     "HI": "ನಮಸ್ಕಾರ, {{ಹೆಸರು}}",
     "MOBILE": "ಮೊಬೈಲ್",
     "ALL_BLOCKS": "ಎಲ್ಲಾ ಬ್ಲಾಕ್‌ಗಳು",
