+++ apps/admin-app-repo/public/locales/mr/common.json
@@ -94,6 +94,7 @@
     "NO_DISTRICTS": "कोई जिले नहीं",
     "ALL_STATES": "सभी राज्य",
     "LOGOUT": "लॉग आउट",
+    "SWITCH_PROGRAM_ROLE": "प्रवेश बदला",
     "HI": "नमस्ते, {{name}}",
     "MOBILE": "मोबाइल",
     "ALL_BLOCKS": "सभी ब्लॉक",
