+++ apps/admin-app-repo/public/locales/odi/common.json
@@ -94,6 +94,7 @@
     "NO_DISTRICTS": "ଜିଲ୍ଲା ନାହିଁ",
     "ALL_STATES": "ସମସ୍ତ ରାଜ୍ୟ",
     "LOGOUT": "ଲଗ୍ ଆଉଟ୍",
+    "SWITCH_PROGRAM_ROLE": "ପ୍ରବେଶ ବଦଳାନ୍ତୁ",
     "HI": "ନମସ୍କାର, {{name}}",
     "MOBILE": "ମୋବାଇଲ୍",
     "ALL_BLOCKS": "ସମସ୍ତ ବ୍ଲକ୍",
