+++ apps/admin-app-repo/public/locales/tam/common.json
@@ -94,6 +94,7 @@
     "NO_DISTRICTS": "மாவட்டங்கள் இல்லை",
     "ALL_STATES": "அனைத்து மாநிலங்கள்",
     "LOGOUT": "வெளியேறு",
+    "SWITCH_PROGRAM_ROLE": "அணுகலை மாற்று",
     "HI": "வணக்கம், {{name}}",
     "MOBILE": "கைபேசி எண்",
     "ALL_BLOCKS": "அனைத்து வட்டங்கள்",
