+++ apps/admin-app-repo/public/locales/tel/common.json
@@ -94,6 +94,7 @@
     "NO_DISTRICTS": "జిల్లాలు లేవు",
     "ALL_STATES": "అన్ని రాష్ట్రాలు",
     "LOGOUT": "లాగ్ అవుట్ చేయండి",
+    "SWITCH_PROGRAM_ROLE": "యాక్సెస్‌ను మార్చండి",
     "HI": "హాయ్, {{పేరు}}",
     "MOBILE": "మొబైల్",
     "ALL_BLOCKS": "అన్ని బ్లాక్‌లు",
