+++ apps/admin-app-repo/public/locales/ur/common.json
@@ -94,6 +94,7 @@
     "NO_DISTRICTS": "کوئی اضلاع نہیں۔",
     "ALL_STATES": "تمام ریاستیں۔",
     "LOGOUT": "لاگ آؤٹ کریں۔",
+    "SWITCH_PROGRAM_ROLE": "رسائی تبدیل کریں",
     "HI": "ہیلو، {{name}}",
     "MOBILE": "موبائل",
     "ALL_BLOCKS": "تمام بلاکس",
